Clean up and fix errors
